+++ modules/tracktion_engine/control_surfaces/types/tracktion_MackieMCU.cpp
@@ -350,7 +350,7 @@ void MackieMCU::acceptMidiMessageInt (int deviceIndex, const juce::MidiMessage& 
     else if (d[0] >= 0xe0 && d[0] <= 0xe8)
     {
         // fader
-        const float pos = juce::jlimit (0.0f, 1.0f, ((((int)d[2]) << 7) + d[1]) * (1.0f / 0x3f70));
+        const float pos = juce::jlimit (0.0f, 1.0f, static_cast<float> (((((int)d[2]) << 7) + d[1])) * (1.0f / 0x3f70));
 
         // send it back to the MCU
         sendMidiCommandToController (deviceIndex, d, 3);
