+++ modules/tracktion_engine/plugins/tracktion_Plugins.test.cpp
@@ -69,7 +69,7 @@ TEST_SUITE ("tracktion_engine")
                 if (transient)
                 {
                     MESSAGE ("Transient diff frames: " << std::to_string (std::abs (transient->first - expectedTransientSample))); // 5 sample tolerance
-                    MESSAGE (std::to_string (transient->first > 0.5f));
+                    MESSAGE (std::to_string (static_cast<float> (transient->first) > 0.5f));
                 }
             }
         };
